requirements.txt
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
dpcpp-cpp-rt==2025.0.4
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fsspec==2024.6.1
idna==3.10
impi-devel==2021.14.1
impi-rt==2021.14.1
intel-cmplr-lib-rt==2025.0.4
intel-cmplr-lib-ur==2025.0.4
intel-cmplr-lic-rt==2025.0.4
intel-extension-for-pytorch==2.7.10+xpu
intel-opencl-rt==2025.0.4
intel-openmp==2025.0.4
intel-pti==0.10.1
intel-sycl-rt==2025.0.4
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.4
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-pygments==0.3.0
markupsafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.1.3
mkl==2025.0.1
mkl-dpcpp==2025.0.1
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.3
numpy==2.1.2
oneccl==2021.14.1
oneccl-bind-pt==2.7.0+xpu
oneccl-devel==2021.14.1
onemkl-sycl-blas==2025.0.1
onemkl-sycl-datafitting==2025.0.1
onemkl-sycl-dft==2025.0.1
onemkl-sycl-lapack==2025.0.1
onemkl-sycl-rng==2025.0.1
onemkl-sycl-sparse==2025.0.1
onemkl-sycl-stats==2025.0.1
onemkl-sycl-vm==2025.0.1
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
python-dateutil==2.9.0.post0
pytorch-triton-xpu==3.3.0
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
ruamel-yaml==0.18.12
ruamel-yaml-clib==0.2.12
setuptools==70.2.0
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.3
tbb==2022.1.0
tcmlib==1.2.0
tinycss2==1.4.0
torch==2.7.0+xpu
torchaudio==2.7.0+xpu
torchvision==0.22.0+xpu
tornado==6.5.1
traitlets==5.14.3
typing-extensions==4.12.2
umf==0.9.1
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
